openai + pipecat + streaming/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
loguru>=0.7.0

# AI and Audio
openai>=1.3.0
gtts>=2.4.0

# Optional dependencies for advanced features
pipecat-ai>=0.0.1a0
torch>=2.0.0
whisper>=1.1.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0